tests: Move `GET /crates/:name/versions` test to `krate::versions` module
